Refactor to remove link for issue #154
