+++ src/main/java/ch/powerunit/extensions/matchers/provideprocessor/ProvidesMatchersAnnotatedElementMatcherMirror.java
@@ -201,5 +201,4 @@ public abstract class ProvidesMatchersAnnotatedElementMatcherMirror
 				getSimpleNameOfGeneratedInterfaceMatcherWithGenericParent(),
 				getFullyQualifiedNameOfClassAnnotatedWithProvideMatcherWithGeneric());
 	}
-
 }
