+++ src/main/java/ch/powerunit/extensions/matchers/provideprocessor/xml/GeneratedMatchers.java
@@ -45,7 +45,7 @@ public class GeneratedMatchers {
 	}
 
 	public Element[] listElements() {
-		return generatedMatcher.stream().map(g -> g.getMirror().getElement()).collect(toList()).toArray(new Element[0]);
+		return generatedMatcher.stream().map(g -> g.getElement()).collect(toList()).toArray(new Element[0]);
 	}
 
 }
